Commit agregar temas materias y cursos

+++ src/main/java/com/singapure/app/models/Cursos.java
@@ -28,7 +28,7 @@ public class Cursos implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="id_curso")
-	private Integer idCurso;
+	private Long idCurso;
 
 	@Column(name="fecha_act")
 	private Date fechaActualizacion;
